Creates a working mobile side navigation bar by adding a context provider within the main container that handles the logic for opening and closing the mobile sidebar.
