+++ src/client/App.jsx
@@ -3,7 +3,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 const queryClient = new QueryClient();
 
-import MainContainer from './Containers/MainContainer.jsx';
+import { MainContainer } from './Containers/MainContainer.jsx';
 
 // App renders from index.js
 const App = () => {
